+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add semantic labels to operator keys and completion overlay
+- Add accessibility semantics for screen readers
 
 
 [0.9.2] - 2026-07-03